CP-1494 #message Adding JSONAPI specifications and Pagination to Contents endpoints

+++ app/Http/Controllers/ContentsController.php
@@ -4,11 +4,12 @@ namespace WA\Http\Controllers;
 
 
 use WA\DataStore\Content\ContentTransformer;
+use WA\DataStore\Content\Content;
 use WA\Http\Controllers\Auth\AuthorizedController;
 use WA\Repositories\Content\ContentInterface;
 use Illuminate\Http\Request;
 
-class ContentsController extends AuthorizedController
+class ContentsController extends ApiController
 {
     /**
      * @var ContentInterface
@@ -32,8 +33,14 @@ class ContentsController extends AuthorizedController
      */
     public function index()
     {
-        $contents = $this->contents->getAllContents();
-        return $this->response()->collection($contents, new ContentTransformer(), ['key' => 'contents']);
+        $criteria = $this->getRequestCriteria();
+        $this->contents->setCriteria($criteria);
+        $contents = $this->contents->byPage();
+
+        $response = $this->response()->withPaginator($contents, new ContentTransformer(),['key'=>'contents']);
+        $response = $this->applyMeta($response);
+        return $response;
+
     }
 
     /**
@@ -44,7 +51,12 @@ class ContentsController extends AuthorizedController
      */
     public function show($id)
     {
-        $content = $this->contents->byId($id);
+        $content = Content::find($id);
+        if(!isset($content))
+        {
+            $error['errors']['put'] = 'Content selected does not exist';
+            return response()->json($error)->setStatusCode(404);
+        }
         return $this->response()->item($content, new ContentTransformer(), ['key' => 'contents']);
     }
 
@@ -56,9 +68,20 @@ class ContentsController extends AuthorizedController
      */
     public function store($id, Request $request)
     {
+        $content = Content::find($id);
+        if(!isset($content))
+        {
+            $error['errors']['put'] = 'Content selected does not exist';
+            return response()->json($error)->setStatusCode(404);
+        }
         $data = $request->all();
         $data['id'] = $id;
         $content = $this->contents->update($data);
+        if(!$content)
+        {
+            $error['errors']['put'] = 'Content could not be updated. Please check your data';
+            return response()->json($error)->setStatusCode(403);
+        }
         return $this->response()->item($content, new ContentTransformer(), ['key' => 'contents']);
     }
 
@@ -71,6 +94,11 @@ class ContentsController extends AuthorizedController
     {
         $data = $request->all();
         $content = $this->contents->create($data);
+        if(!$content)
+        {
+            $error['errors']['post'] = 'Content could not be created. Please check your data';
+            return response()->json($error)->setStatusCode(403);
+        }
         return $this->response()->item($content, new ContentTransformer(), ['key' => 'contents']);
     }
 
@@ -78,10 +106,25 @@ class ContentsController extends AuthorizedController
      * Delete content
      *
      * @param $id
+     * @return \Symfony\Component\HttpFoundation\Response
      */
     public function deleteContent($id)
     {
+        $content = Content::find($id);
+        if(!isset($content))
+        {
+            $error['errors']['delete'] = 'Content selected does not exist';
+            return response()->json($error)->setStatusCode(404);
+        }
+
         $this->contents->deleteById($id);
+        $content = Content::find($id);
+        if(!$content)
+        {
+            return response()->json()->setStatusCode(204);
+        }else{
+            return response()->json()->setStatusCode(202);
+        }
         $this->index();
     }
